+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 All notable changes to the Viewer Metrics Chrome Extension will be documented in this file.
 
-## [0.9.82] - 2025-11-24
-### Fixed
+## [0.9.9] - 2025-11-29
+- New High Churn mode for channels with short-lived bots
+- Re-worked the stats display to be more intuitive
 - Smooth lines now an easy toggle from main chart
-- User retention configurable
+- User retention and summary start configurable
 - Bot retention separated from viewer retention
 - Outgoing requests are now paused on stream end via auto-pause option
+- Can now skip the first x minutes of main graph where calculations are still taking place
+- Time buckets fixed and now includes all viewers seen
 
 ## [0.9.81] - 2025-11-24
 ### Fixed